modified type to Long as there's a conflict in the controller

+++ src/main/java/team/nine/Exams/repositories/UserRepository.java
@@ -6,7 +6,7 @@ import team.nine.Exams.models.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     @Query(value = "SELECT u from User u WHERE u.userName = ?1")
     Optional<User> findUserName(String username);
 
@@ -18,7 +18,6 @@ public interface UserRepository extends JpaRepository<User,Integer> {
 
     @Query(value = "SELECT u from User u WHERE u.token = ?1")
     Optional<User> findByToken(String token);
-
 
 
     boolean existsByUserName(String username);
